HAJ-1: update validation and types

+++ shj-admin-portal-orm/src/main/java/com/elm/shj/admin/portal/orm/entity/JpaCompanyLite.java
@@ -43,7 +43,7 @@ public class JpaCompanyLite implements Serializable {
     @Column(name = "type_code")
     private String typeCode;
     @Column(name = "mission_ref_code")
-    private Integer missionRefCode;
+    private Long missionRefCode;
     @Column(name = "establishment_ref_code")
     private Integer establishmentRefCode;
 }
