+++ shj-admin-portal-services/src/main/java/com/elm/shj/admin/portal/services/dto/EHousingType.java
@@ -19,17 +19,17 @@ public enum EHousingType {
     CAMP(1),
     HOTEL(2),
     BUILDING(3);
-    private final long id;
+    private final int id;
 
-    EHousingType(long id) {
+    EHousingType(int id) {
         this.id = id;
     }
 
-    public long getId() {
+    public int getId() {
         return this.id;
     }
 
-    public static EHousingType fromId(long id) {
+    public static EHousingType fromId(int id) {
         List<EHousingType> result = Arrays.stream(EHousingType.values()).filter(e -> e.getId() == id).collect(Collectors.toList());
         return result.isEmpty() ? null : (EHousingType) CollectionUtils.get(result, 0);
     }
